// src/pages/evaluation/schema.ts
import { z } from "zod"

export const perfomanceEvaluationSchema = z.object({
    id: z.string(),
    number: z.string(),
    contractNumber: z.string(),
    vendor: z.string(),
    contractType: z.string(),
    contractStartDate: z.string(),
    contractEndDate: z.string(),
    initialContractDuration: z.string(),
    finalContractDuration: z.string(),
    initialContractSum: z.string(),
    finalContractSum: z.string(),
    contractManager: z.string(),
    description: z.string(),
    numberOfAddendums: z.string(),
    // vendor evaluations
    capacityScore: z.number(),
    capicityComment: z.string(),
    competencyScore: z.number(),
    competencyComment: z.string(),
    costManagementScore: z.number(),
    costManagementComment: z.string(),
    qualityScore: z.number(),
    qualityComment: z.string(),
    qualityConsistency: z.number(),
    qualityConsistencyComment: z.string(),
    environmentConsiderationScore: z.number(),
    environmentConsiderationComment: z.string(),
    communicationScore: z.number(),
    communicationComment: z.number(),
    financialCapabilityScore: z.number(),
    financialCapabilityComment: z.number(),
    overallScore: z.number(),
    implemantionPercentage: z.number(),
    timeUsedPercentage: z.number(),
    statusCode: z.string(),
    statusName: z.string(),
  })

  export const perfomanceEvaluationFormSchema = z.object({
    capacityScore: z.number(),
    capicityComment: z.string(),
    competencyScore: z.number(),
    competencyComment: z.string(),
    costManagementScore: z.number(),
    costManagementComment: z.string(),
    qualityScore: z.number(),
    qualityComment: z.string(),
    qualityConsistency: z.number(),
    qualityConsistencyComment: z.string(),
    environmentConsiderationScore: z.number(),
    environmentConsiderationComment: z.string(),
    communicationScore: z.number(),
    communicationComment: z.number(),
    financialCapabilityScore: z.number(),
    financialCapabilityComment: z.number(),
    overallScore: z.number(),
    implemantionPercentage: z.number(),
    timeUsedPercentage: z.number(),
  })
  export type PerfomanceEvaluation = z.infer<typeof perfomanceEvaluationSchema>